Add files via upload
These are the codes needed to prepare a MH-enabled run. One sub-directory contains the codes for creating the Green's functions which the code uses to calculate the LW field, the other sub-directory contains the codes to calculate the 'morphed' densities, i.e. density grids interpolated to finer timesteps than the original ones (typically 2 Myr, comparable to PopIII star lifetime).